+++ packages/redux/src/reducers/entities.js
@@ -4,20 +4,14 @@ import types from '../constants/action-types'
 
 // lodash
 import concat from 'lodash/concat'
-import filter from 'lodash/filter'
 import forEach from 'lodash/forEach'
 import get from 'lodash/get'
-import map from 'lodash/map'
-import merge from 'lodash/merge'
 import values from 'lodash/values'
 
 const _ = {
   concat,
-  filter,
   forEach,
   get,
-  map,
-  merge,
   values,
 }
 
@@ -86,16 +80,8 @@ function _buildState(allIds, entities, overwriteExisted = false) {
 function entities(state = defaultState, action = {}) {
   switch (action.type) {
     case types.indexPage.read.success: {
-      const allPostIds = _.get(
-        state,
-        [fieldNames.postsInEntities, 'allIds'],
-        []
-      )
-      const allTopicIds = _.get(
-        state,
-        [fieldNames.topicsInEntities, 'allIds'],
-        []
-      )
+      const allPostIds = state.posts.allIds
+      const allTopicIds = state.topics.allIds
 
       const fieldKeys = _.concat(
         _.values(fieldNames.sections),
@@ -117,85 +103,126 @@ function entities(state = defaultState, action = {}) {
         }
       })
 
-      return _.merge({}, state, {
-        [fieldNames.postsInEntities]: _buildState(allPostIds, posts),
-        [fieldNames.topicsInEntities]: _buildState(allTopicIds, topics),
-      })
+      const newStateForPosts = _buildState(allPostIds, posts)
+      const newStateForTopics = _buildState(allTopicIds, topics)
+
+      return {
+        posts: {
+          allIds: newStateForPosts.allIds,
+          byId: Object.assign({}, state.posts.byId, newStateForPosts.byId),
+          slugToId: Object.assign(
+            {},
+            state.posts.slugToId,
+            newStateForPosts.slugToId
+          ),
+        },
+        topics: {
+          allIds: newStateForTopics.allIds,
+          byId: Object.assign({}, state.topics.byId, newStateForTopics.byId),
+          slugToId: Object.assign(
+            {},
+            state.topics.slugToId,
+            newStateForTopics.slugToId
+          ),
+        },
+      }
     }
 
     case types.topics.read.success: {
-      const allTopicIds = _.get(
-        state,
-        [fieldNames.topicsInEntities, 'allIds'],
-        []
-      )
+      const allTopicIds = state.topics.allIds
       const topics = _.get(action, 'payload.items', [])
 
-      return _.merge({}, state, {
-        [fieldNames.topicsInEntities]: _buildState(allTopicIds, topics),
-      })
+      const { allIds, byId, slugToId } = _buildState(allTopicIds, topics)
+
+      return {
+        ...state,
+        topics: {
+          allIds,
+          byId: Object.assign({}, state.topics.byId, byId),
+          slugToId: Object.assign({}, state.topics.slugToId, slugToId),
+        },
+      }
     }
 
     case types.selectedPost.read.success: {
-      const allPostIds = _.get(
-        state,
-        [fieldNames.postsInEntities, 'allIds'],
-        []
-      )
+      const allPostIds = state.posts.allIds
       const post = _.get(action, 'payload.post', {})
 
-      return _.merge({}, state, {
-        [fieldNames.postsInEntities]: _buildState(allPostIds, post, true),
-      })
+      const { allIds, byId, slugToId } = _buildState(allPostIds, post, true)
+
+      return {
+        ...state,
+        posts: {
+          allIds,
+          byId: Object.assign({}, state.posts.byId, byId),
+          slugToId: Object.assign({}, state.posts.slugToId, slugToId),
+        },
+      }
     }
 
     case types.selectedTopic.read.success: {
-      const allTopicIds = _.get(
-        state,
-        [fieldNames.topicsInEntities, 'allIds'],
-        []
-      )
+      const allTopicIds = state.topics.allIds
       const topic = _.get(action, 'payload.topic', {})
 
-      return _.merge({}, state, {
-        [fieldNames.topicsInEntities]: _buildState(allTopicIds, topic, true),
-      })
+      const { allIds, byId, slugToId } = _buildState(allTopicIds, topic, true)
+
+      return {
+        ...state,
+        topics: {
+          allIds,
+          byId: Object.assign({}, state.topics.byId, byId),
+          slugToId: Object.assign({}, state.topics.slugToId, slugToId),
+        },
+      }
     }
 
     case types.postsByListId.read.success:
     case types.relatedPosts.read.success: {
-      const allPostIds = _.get(
-        state,
-        [fieldNames.postsInEntities, 'allIds'],
-        []
-      )
+      const allPostIds = state.posts.allIds
       const posts = _.get(action, 'payload.items', [])
 
-      return _.merge({}, state, {
-        [fieldNames.postsInEntities]: _buildState(allPostIds, posts),
-      })
+      const { allIds, byId, slugToId } = _buildState(allPostIds, posts)
+
+      return {
+        ...state,
+        posts: {
+          allIds,
+          byId: Object.assign({}, state.posts.byId, byId),
+          slugToId: Object.assign({}, state.posts.slugToId, slugToId),
+        },
+      }
     }
 
     case types.featureTopic.read.success: {
-      const allPostIds = _.get(
-        state,
-        [fieldNames.postsInEntities, 'allIds'],
-        []
-      )
-
-      const allTopicIds = _.get(
-        state,
-        [fieldNames.topicsInEntities, 'allIds'],
-        []
-      )
+      const allPostIds = state.posts.allIds
+      const allTopicIds = state.topics.allIds
 
       const posts = _.get(action, 'payload.lastThreeRelatedPosts', [])
       const topics = [_.get(action, 'payload.topic', {})]
 
-      return _.merge({}, state, {
-        [fieldNames.postsInEntities]: _buildState(allPostIds, posts),
-        [fieldNames.topicsInEntities]: _buildState(allTopicIds, topics),
-      })
+      const newStateForPosts = _buildState(allPostIds, posts)
+      const newStateForTopics = _buildState(allTopicIds, topics)
+
+      return {
+        posts: {
+          allIds: newStateForPosts.allIds,
+          byId: Object.assign({}, state.posts.byId, newStateForPosts.byId),
+          slugToId: Object.assign(
+            {},
+            state.posts.slugToId,
+            newStateForPosts.slugToId
+          ),
+        },
+        topics: {
+          allIds: newStateForTopics.allIds,
+          byId: Object.assign({}, state.topics.byId, newStateForTopics.byId),
+          slugToId: Object.assign(
+            {},
+            state.topics.slugToId,
+            newStateForTopics.slugToId
+          ),
+        },
+      }
     }
 
     default: {