+++ packages/react-article-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.6-rc.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-article-components@1.2.5...@twreporter/react-article-components@1.2.6-rc.0) (2020-09-15)
+
+**Note:** Version bump only for package @twreporter/react-article-components
+
+
+
+
+
 ## [1.2.5](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-article-components@1.2.5-rc.0...@twreporter/react-article-components@1.2.5) (2020-09-14)
 
 **Note:** Version bump only for package @twreporter/react-article-components